+++ link_test.go
@@ -235,7 +235,7 @@ func newTestLink(t *testing.T) *Receiver {
 			source: &frames.Source{},
 			// adding just enough so the debug() print will still work...
 			// debug(1, "FLOW Link Mux half: source: %s, inflight: %d, credit: %d, deliveryCount: %d, messages: %d, unsettled: %d, maxCredit : %d, settleMode: %s", l.source.Address, l.receiver.inFlight.len(), l.l.availableCredit, l.deliveryCount, len(l.messages), l.countUnsettled(), l.receiver.maxCredit, l.receiverSettleMode.String())
-			detached: make(chan struct{}),
+			done: make(chan struct{}),
 			session: &Session{
 				tx:   make(chan frames.FrameBody, 100),
 				done: make(chan struct{}),